rename push_boulder to boulder
